LPS-198743 Extract DisplayPageSelector component to a separate file in commons folder so that it can be reused
