+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/DisplayPageSelector.js
@@ -1,0 +1,89 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import PropTypes from 'prop-types';
+import React, {useEffect, useState} from 'react';
+
+import {addMappingFields} from '../../app/actions/index';
+import {useDispatch, useSelector} from '../../app/contexts/StoreContext';
+import InfoItemService from '../../app/services/InfoItemService';
+import getMappingFieldsKey from '../../app/utils/getMappingFieldsKey';
+import MappingFieldSelector from './MappingFieldSelector';
+
+function filterFields(fieldSets) {
+	return fieldSets.reduce((acc, fieldSet) => {
+		const newFields = fieldSet.fields.filter(
+			(field) => field.type === 'display-page'
+		);
+
+		if (newFields.length) {
+			return [
+				...acc,
+				{
+					...fieldSet,
+					fields: newFields,
+				},
+			];
+		}
+
+		return acc;
+	}, []);
+}
+
+export default function DisplayPageSelector({
+	classNameId,
+	classTypeId,
+	onConfigChange,
+	selectedValue,
+}) {
+	const dispatch = useDispatch();
+
+	const mappingFields = useSelector((state) => state.mappingFields);
+
+	const [displayPageFields, setDisplayPageFields] = useState(null);
+
+	useEffect(() => {
+		const key = getMappingFieldsKey({classNameId, classTypeId});
+
+		const fieldSets = mappingFields[key];
+
+		if (fieldSets) {
+			setDisplayPageFields(filterFields(fieldSets));
+		}
+		else {
+			InfoItemService.getAvailableStructureMappingFields({
+				classNameId,
+				classTypeId,
+			}).then((newFields) => {
+				dispatch(addMappingFields({fields: newFields, key}));
+			});
+		}
+	}, [dispatch, classNameId, classTypeId, mappingFields]);
+
+	return (
+		<MappingFieldSelector
+			className="mb-3"
+			defaultLabel={`-- ${Liferay.Language.get('none')} --`}
+			fields={displayPageFields}
+			label={Liferay.Language.get('display-page')}
+			onValueSelect={(event) =>
+				onConfigChange({
+					displayPage:
+						event.target.value === 'unmapped'
+							? null
+							: event.target.value,
+				})
+			}
+			value={selectedValue}
+		/>
+	);
+}
+
+DisplayPageSelector.propTypes = {
+	classNameId: PropTypes.string.isRequired,
+	classTypeId: PropTypes.string.isRequired,
+	onConfigChange: PropTypes.func.isRequired,
+	selectedValue: PropTypes.string,
+};
